nix: fix rocksdb, and always use 10.2.1
Seemingly, after recent commits to Mina,
compiling stopped working on non-x86 platforms.

This is probably because on such platforms, the code would default
to a newer rocksdb from Nixpkgs, which now (?) tries to use liburing
by default. liburing is not used by Mina.

The fix is to compile rocksdb (as produced by Nixpkgs) without liburing support.
Some other clean-ups have also been done at the same time,
and since seemingly there was no point in using the ancient version of rocksdb
on x86, 10.2.1 is now used there too.

This aligns the version of rocksdb used with Nix with the version of rocksdb
that is likely to be used without Nix in the near future AFAICT.